Updates logic to get human choice and play round
Extracts scoreRound from the playGame function to the global space.
Extracts part of playRound from playGame to global space. Ultimately,
the playGame function will be deleted as the game has been changed from
5 rounds to first to 5. I probably should have put these in the global
space to begin with.

Adds event listener to the button panel to determine which button was
pressed. This is slightly more efficient than adding an event listener
for each button because there is less setup for the browser to do.

Refactors code to use uppercase for string constants.