Refactor how buttons are hidden/shown
Updates how buttons are hidden/shown to use the display style attribute
instead of the hidden property. hidden should be used when you want to
hide an element but still reserve its space in the layout whereas
display: none should be used when you want to completely remove an
element from the page, both visually and in terms of layout space.